migrations/models/1_20251018150222_allow_officer_approval_null.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "officers" ALTER COLUMN "approval_by" DROP NOT NULL;
    """


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "officers" ALTER COLUMN "approval_by" SET NOT NULL;
    """
