requirements.txt
aerich==0.7.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.29.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==43.0.3
dictdiffer==0.9.0
fastapi==0.115.12
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
requests==2.32.3
iso8601==2.1.0
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.10.1
pydantic_core==2.33.0
pydantic[email]>=2.5
email-validator>=2.1.0.post1
PyJWT[crypto]==2.10.1
pypika-tortoise==0.6.1
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
tomlkit==0.13.3
tortoise-orm==0.25.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
redis>=5.0.0
uvicorn==0.32.1
watchfiles==1.1.0
websockets==15.0.1
pytest==8.3.3
aiomysql==0.2.0
PyMySQL==1.1.1